+++ prisma/migrations/20260312020026_add_soft_delete_and_dispute_email/migration.sql
@@ -1,5 +1,5 @@
 -- AlterTable
-ALTER TABLE "Estate" ADD COLUMN "deletedAt" TIMESTAMP(3);
+ALTER TABLE "Estate" ADD COLUMN     "deletedAt" TIMESTAMP(3);
 
 -- AlterTable
-ALTER TABLE "SiteSettings" ADD COLUMN "disputeEmail" TEXT NOT NULL DEFAULT 'dispute@eorzeaestates.com';
+ALTER TABLE "SiteSettings" ADD COLUMN     "disputeEmail" TEXT NOT NULL DEFAULT 'dispute@eorzeaestates.com';